@@ -18,22 +18,33 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, this.title}) : super(key: key);
 
-  final String title;
+  final String? title;
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
 
 class _MyHomePageState extends State<MyHomePage> {
+  List<int> items = [];
+
+  @override
+  void initState() {
+    super.initState();
+
+    for (int i = 0; i < 10; i++) {
+      items.add(i);
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     return DefaultTabController(
       length: 2,
       child: Scaffold(
         appBar: AppBar(
-          title: Text(widget.title),
+          title: Text(widget.title!),
           bottom: TabBar(
             tabs: [
               Tab(text: "Tab1"),
@@ -61,9 +72,14 @@ class _MyHomePageState extends State<MyHomePage> {
                 onWeightChanged: (w) => print("Horizon: $w"),
               ),
               view2: Container(
-                child: Center(
-                  child: Text("View3"),
-                ),
+                child: ListView.separated(
+                    itemBuilder: (context, index) {
+                      return ListTile(
+                        title: Text("Item ${items[index]}"),
+                      );
+                    },
+                    separatorBuilder: (context, index) => Divider(),
+                    itemCount: items.length),
                 color: Colors.green,
               ),
               viewMode: SplitViewMode.Vertical,
